requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.12.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Database and Caching
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.0.0
pgvector>=0.1.0

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.0
python-magic>=0.4.24
numpy>=1.24.3
tiktoken>=0.3.0
pdf2image>=1.16.3  # Replaces poppler-utils as Python package

# AI and Embeddings
openai>=1.0.0
scikit-learn>=0.24.2
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
spacy>=3.5.0

# Monitoring and Logging
prometheus-client>=0.12.0
python-json-logger>=2.0.0
structlog>=21.1.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
httpx>=0.18.2

# Utilities
python-dotenv>=0.19.0
requests>=2.26.0
aiofiles>=0.7.0
tenacity>=8.0.1
python-dateutil>=2.8.2
beautifulsoup4>=4.9.3
lxml>=4.9.0

# Development
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910

# Streamlit Components
streamlit-chat>=0.0.2.1
streamlit-aggrid>=0.3.3
streamlit-ace>=0.1.1

# Performance Optimization
ujson>=5.4.0
orjson>=3.8.0
uvloop>=0.17.0
httptools>=0.5.0
